Created a package structure. Moved to example data - CSP50compare - CSPCret - CSPecon
