+++ .gitignore
@@ -2,3 +2,5 @@
 /tex/
 __pycache__/
 
+/CombiCSP.egg-info/
+/doc/tex/
