+++ examples/example_data/tmy_35.015_25.755_2005_2020.csv
